split exchange and correlation in separate IO
Because exchange and correlation are calculated by different parts of the program
it is easier to split the writing routine in two separate routines.

The disadvantage of this approach is that the IO routine becomes aware of the
structure of the code. Hence, we must replace the order in the reading routine.

use ikpt instead of iunit (otherwise all tags would be identical and
there would be no way to know the tag)