introduce new linear_op function
The linear operator (H - w S + a_pv P_v) is used in several parts of the code.
Because they calls are slightly different, a separate routine was written for
every one of these calls. This is not very efficient, so that we introduce the
new file linear_op that should be able to handle all of these calls.