make more use of new parallel_task in do_stern
The old implementation used an mp_sum to get all information to the root process
which does the post-processing of the matrix data. This is inefficient for two
reasons: i) the mp_sum call collects the result on all processes, even though only
one uses it; ii) the mp_sum adds a lot of zeros which results in a lot of
communication overhead.

The new implementation uses a call to mp_gatherv instead so that the result is
only gathered on the single process that actually uses it. To facilitate this,
we needed to do a significant rewrite of parts of the code.

The arrays allocated on the other processes now only have the dimension that is
actually used. The order of the indices is changes, so that the slowest running
index is the last one, which is allows for more efficient contigous memory access.
This directly affected coulomb, which fills these arrays. As a result the unfold_w
routine was changed to do both reordering of the indices as well as the symmetry
unfolding.

In addition, we cleaned up do_stern significantly to be consistent with the code
style and remove any unused variables.