implement the BICG part of Fromme's paper
The implementation follows the paper closely. With the comments, we indicate
which line of the algorithm the operation corresponds to. We use BLAS operations
to do all the vector transformations to optimize the speed of the code.

The additional variables are stored in the type (even though they are not
necessary there) to provide a consistent interface, e.g., all variables
associated with the shifted systems are accessed via the shifted_system type.