split solver for first iteration
In order to unify the solvers the first iteration of the iterative solver is
split of from the rest so that we may use a different solver there. It is
planned to replace the first iteration solver with the multishift solver, but
for now the solver is the same as for the later iterations.

In addition, we introduced a frequency dimension, so that we can iterate over
multiple frequencies later similar like the phonon code iterates over multiple
perturbations. This logic follows already the idea that the multishift will
be used for the first iteration. Currently, the number of frequencies is just
set to 1, so that every frequency is calculated independently.

In addition, we did some more code cleanup and added consistency to the code style.