MAJOR change in frequency layout
We use the new frequency type throughout the code. In particular, we replace
all use statements of freq_gw that refer to the arrays wcoul, wgreen, wsigma,
wgtcoul, or wgtcoul with references to the new type. The size of the arrays
are still passed with the module, though.

Because the new type stores the frequencies in Ry, gwq_readin needs to convert
the input frequencies to Ry. We also modify the routines that use the frequencies
and eliminate the conversion to Ry.

We replace the convergence in the step size delta with an equivalent one in
the number of steps.