wrap fft6 in module
There are currently three different implementations of fft6 (with or without
symmetry). The plan is to collect all of these into a single module that is
then used by all of the routines.

In this commit, we create a module wrapping the fft6 routine and link the files
using this routine to the new module.