split setup_nscf in two files
For W, there are two grids k and q similar to phonon. We will obtain W on the q
grid by evaluating perturbations at all k + q. The routine setup_nscf does this
and should be replaced by the one LR_Modules.

We want to evaluate Sigma at a user defined mesh of k-points, for which we need
the Green's function at all points k - q. The q points are the ones where W is
defined. The present routine does not work properly for this purpose, because
it generates a new k-point mesh that is not related to the user defined one.