MAJOR change of setup_nscf_green
Instead of following strictly the setup of the phonon routine, we write a
different setup routine that is much closer to the task that we want to do.
To evaluate Sigma_k = sum_S G_{k-Sq} W_{Sq}, we construct for every k point
a mesh that contains k and all k - Sq and then perform non self-consistent
calculations for all of these points.

We adjust the ikks and ikqs maps appropriately by passing an additional argument
that identifies whether we are in the coulomb or self-energy part of the code.

NOTE: the results of the calculation change, because the Fermi energy of the
      new grid might be different from the old one
NOTE: the exchange is currently broken, because it wasn't adjusted to the new
      grid yet