create new green module
The plan is to unify the calculation of the Green's function independent if the
frequencies are real or complex into a single routine. For this we create a new
module in the file green.f90.

So far this module only contains a dummy subroutine that checks the sanity of the
input and the module is added to the makefile.