bugfix for iterative solver
The parallelization introduced a bug to the iterative solver. A different
array was introduced which contains the screened Coulomb interaction which
was then copied into the scrcoul_g array written to file. In this process
the |q + G| = 0 term got lost, i.e. was overwritten by the copied data.

Hence, we move the copy of the diagonal term after the unfold/copy instruction.
We also clean-up the equivalent copy instruction in invert_epsilon, which
also reduces the API.